# docs/swagger.yaml
info:
  contact: {}
  license: {}
paths:
...
  /api/rests/addMyRest:
    post:
      parameters:
      - description: 申请调休时间
        in: query
        name: request_time
        required: true
        type: string
      - description: 申请调休类型，0：上午，1：下午，2：全天
        in: query
        name: type
        required: true
        type: integer
      - description: 申请调休类型，0：调休，1：年休
        in: query
        name: vacation_type
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 新增本人调休申请表信息
...
  /api/rests/deleteMyRest{id}:
    post:
      parameters:
      - description: Id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 删除本人未审批调休申请表信息
  /api/rests/examineRest:
    post:
      parameters:
      - description: Id
        in: query
        name: id
        required: true
        type: integer
      - description: 回复，状态 0为默认、1为拒绝、2为同意
        in: query
        name: response
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 审批调休申请表
...
  /api/rests/getMe:
    post:
      parameters:
      - description: ' 0 未审批、1 已审批'
        in: query
        name: state
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{rest},"msg":"ok"}'
          schema:
            type: string
      summary: 获取本人调休申请表信息(未审批/已审批)
...
  /api/rota/import:
    post:
      parameters:
      - description: excel表格文件
        in: query
        name: file
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 导入值班表
  /api/rotas/addDay:
    post:
      parameters:
      - description: 日期
        in: query
        name: datetime
        required: true
        type: string
      - description: 星期
        in: query
        name: week
        required: true
        type: string
      - description: 计费晚班人员
        in: query
        name: billing_late
        required: true
        type: string
      - description: 计费周末白班人员
        in: query
        name: billing_weekend_day
        required: true
        type: string
      - description: crm晚班人员
        in: query
        name: crm_late
        required: true
        type: string
      - description: crm周末白班人员
        in: query
        name: crm_weekend_day
        required: true
        type: string
      - description: crm工作日特殊班值班人员
        in: query
        name: crm_duty_special
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 添加日值班表
  /api/rotas/deleteDay{datetime}:
    post:
      parameters:
      - description: 日期，例如：2020-09-01
        in: query
        name: datetime
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 删除日值班表
  /api/rotas/deleteMonth{month}:
    post:
      parameters:
      - description: 月份，例如：2020-09
        in: query
        name: month
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 删除月值班表
  /api/rotas/getMonth{month}:
    post:
      parameters:
      - description: 月份，例如：2020-09
        in: query
        name: month
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取当月值班表
...
  /api/vacation/deleteByName:
    post:
      parameters:
      - description: 姓名
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 删除某人调休信息
  /api/vacation/editByName:
    post:
      parameters:
      - description: 姓名
        in: query
        name: name
        required: true
        type: integer
      - description: 剩余调休
        in: query
        name: remain_vacation
        required: true
        type: number
      - description: 剩余年休
        in: query
        name: remain_annual_vacation
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 修改某人调休信息
...
swagger: "2.0"